Python 2 compatibility
Use `HTMLParser` (Python2) if `html` (Python 3) fails to import.
Use `oscpy` instead of `python-osc` as it is Python 2 and 3 compatible.
It takes bytes instead of strings, so need to do a bit of utf8 en/de-coding to make it work,
and add a sleep loop as it runs the server itself in another thread.